requirements.txt
# Divine Tribe Email Assistant Dependencies

# WooCommerce API
woocommerce==3.0.0

# Gmail API
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Telegram Bot
python-telegram-bot==20.7

# Claude AI
anthropic==0.39.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1

# Flask (for optional web dashboard)
flask==3.0.0
flask-cors==4.0.0
